requirements.txt
ftfy
regex
tqdm
scikit-learn
jittor==1.3.8.4
numpy==1.24.4
pillow==9.5.0